% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_summary_fxns.r
\name{broodsummary.hpd}
\alias{broodsummary.hpd}
\title{broodsummary.hpd.}
\usage{
broodsummary.hpd(x, broodsize, S, D)
}
\description{
internal functionto calculate HPD and append true data-generating parmeters to the results
of simulations from JAGS.
}
