% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_summary_fxns.r
\name{process.simfile}
\alias{process.simfile}
\title{process.simfile.}
\usage{
process.simfile(results_path, results_pattern, sim_broodsize, sim_S, sim_D,
  sim_prior)
}
\arguments{
\item{results_path}{path to the folder containing .rds files .}

\item{results_pattern}{regex expression to pull subset of .rds files based on naming pattern.}

\item{sim_S}{the survival parameter value used in data generation.}

\item{sim_D}{the detection parameter value used in data generation.}

\item{sim_prior}{the type of prior used in the model.}

\item{sim_broodszie}{the number of broods used when simulating the data.}
}
\description{
process.simfile used to read-in and organize the contents of a folder. Folder should contain the output .rds files
from jagsUI.
}
