% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericJAGSsim_fxn.r
\name{repeat.sims}
\alias{repeat.sims}
\title{repeat.sims.
A function to streamline simulations using JAGS in R
runs jagsUI.}
\usage{
repeat.sims(SIMdata.fxn, SIMdata.params, JAGdata.fxn, PTM, JAG.initial,
  chains, thin, iter, burn, adapt, outfolder.path, out_label = "",
  JAGmodel.path, n.sims)
}
\arguments{
\item{SIMdata.fxn}{a    function to simulate your data.}

\item{SIMdata.params}{parameter values passed to your data-generating function.}

\item{JAGdata.fxn}{function to produce a data object to be fed to jags
at minimum, needs at least 1 argument (the simulated data).}

\item{PTM}{vector of parameters to monitor in jags.}

\item{JAG.initial}{function/list passed to jags containing initial values or steps to produce initial values.}

\item{chains}{to jagsUI.}

\item{thin}{to jagsUI.}

\item{iter}{to jagsUI.}

\item{burn}{to jagsUI.}

\item{adapt}{to jagsUI.}

\item{outfolder.path}{path to folder where simulation results should be written.}

\item{JAGmodel.path}{path to location of JAG model file.}

\item{out_Label}{label to add to each rds file.}
}
\description{
repeat.sims.
A function to streamline simulations using JAGS in R
runs jagsUI.
}
